DEVPROD-5156 Move trace uploading to taskoutput package
