+++ config.go
@@ -37,7 +37,7 @@ var (
 
 	// Agent version to control agent rollover. The format is the calendar date
 	// (YYYY-MM-DD).
-	AgentVersion = "2025-03-18"
+	AgentVersion = "2025-03-18a"
 )
 
 const (
